feat: improve architecture
