+++ constants/upload.ts
@@ -1,0 +1,3 @@
+export const MAX_FILE_SIZE = 1024 * 1024; // 1MB
+export const ALLOWED_FILE_TYPES = ['image/jpeg', 'image/png', 'image/svg+xml'];
+export const MAX_IMAGE_DIMENSION = 2048; // píxeles
